Add `CONTRIBUTING.md` file
I'm a new contributor here, and I was looking for this file for a bit and
didn't find one, so I figured I'd add one in. For now I just pulled from the
`contributing` section of the README, but now that this file exists there can
be further instructions (or helpful hints) if needed.