Refactor to remove single pipes
I wanted to do a little refactoring to see how much of a difference the new
single pipe rule would make, and it turns out that a) we were using that
_everywhere_, and b) it actually makes the code a lot easier to reason about.
I think it's a great rule!

This commit doesn't remove _all_ of the single pipe instances since I figured
doing that in one huge PR might be overwhelming, but it's a step in the right
direction. If it would be helpful to have all of them removed in one PR then
I'm happy to do that!