Add UI output for skipped checks
If we're going to skip checks because the user's Elixir version doesn't support
them, I wanted to make sure that the user was as aware of this as possible so
they don't think the check is running but just not reporting any violations.
Since they can explicitly include a check in their configuration and that
configuration would be ignored, I worried that there would be confusion around
that behavior.

Plus, this would be a subtle way to prod people into upgrading to the newest
version of Elixir, which is always a good thing!