Add additional validations
I originally didn't have all the info needed for the validation logic
to make sure a `.credo.exs` file was valid. I think with this commit it
might be there.

I couldn't think of a way to ensure that all the checks were valid
without actually requiring all the files listed in the `requires` part
of the `.credo.exs` file before checking to see if all the listed
modules were compiled, but if there's something better then I'd love
to see it!