Reimplement .credo.exs validation
We needed to re-implement this feature since this way we could have
access to the line numbers on which some of these errors occured.
Because of the way we're parsing the AST in the `Credo.Code` module,
the line numbers are only available on the issue of a check being
incorrect, but frankly that's probably the one that's most needed, so
I'm ok with that.