Add :allow_shorter_alias param to AliasUsage
This adds an option allowing to use a shorter alias than possible, as
this sometimes improves readability.

The AliasUsage check requires to alias used modules so that dependencies
are easier to spot [[1]], for example:

```elixir
defmodule Test do
  alias MyApp.External.MyProto.Client

  def something do
    Client.search(...)
  end
end
```

But sometimes it is more readable to call "MyProto.Client.search"
instead of "Client.search". The dependency can still be seen at the top
of the file (only slightly less specific):

```elixir
defmodule Test do
  alias MyApp.External.MyProto

  def something do
    MyProto.Client.search(...)
  end
end
```

[1]: https://github.com/rrrene/elixir-style-guide#alias-modules