Use call scope in ABC for excluded functions
This commit refactors the ABCSize check.

The ABCSize has an option to ignore any scores when inside calls to
"excluded functions", so that, for example, the Ecto DSL will not
drive the metric upwards.

The previous code removes from the AST anything within a call to
a function marked as excluded, using the AST (under analysis) to
keep track of state in the traversal.

This version pushes the function name to a stack modeling the scope,
and checks the stack for any excluded function when scoring the code,
so the state of the traversal becomes separated from the AST itself.

It is not exactly a problem, but I would love feedback on the result.