fix credo diff cli
When it runs the `mix credo diff` CLI it returns a different value instead of 0.

So it keep the tuple for the pattern matching, but it change the zero value for
an unused variable like `_value`.

And after that, the CLI continues with a weird behavior, because it tells could
not cd to find the path `/tmp/credo-diff-#{now}`, and to guarantee it will create
the path put an if conditions where checks the path it exists if not create a new owns.