Check UnsafeToAtom along pipe chains
Currently, if a call to `String.to_atom/1`, `List.to_atom/1`,
`Module.concat/{1,2}`, `:erlang.list_to_atom/1` and
`:erlang.binary_to_atom/2` is in a pipe (`|>`) chain, it is not
detected by the `UnsafeToAtom` check.

This change alters the prewalk check to unpipe those chains and insert
a phantom argument in the chain to make the arity "correct" and allow
detection.