Add autocorrect to default suggest pipeline
At this point the pipeline for the default "suggest" task has been
modified to include an Autocorrect step. That's all hooked up, and a
callback for `autocorrect/1` has been added to the `Credo.Check`
behaviour and the default implementation. This can be implemented for
each check individually, and it's a noop for now.