Remove autocorrected issues before displaying results
Now when we autocorrect an issue, we remove that issue from the stored
list of issues so it isn't shown to the user (since technically it's no
longer an issue since it was autocorrected).

One other thing we might do is to change the formatter to display when
an issue was there and to indicate that it was autocorrected, also
making sure the exit code is 0 for any autocorrected issues, but for now
this seemed like a reasonable path forward and much simpler than this
option.