"Finish" implementing autocorrect for ordering aliases
We've now handled pretty much all the use cases for ordering aliases
automatically. There is some formatter whackiness that will need to be
figured out, but beyond that this is a good example of how an
autocorrect callback might be implemented for a more involved check than
the first one.