Add autocorrect callback for line endings consistency check
This is the first implementation of an autocorrect callback for a
consistency check. I don't _love_ how we're doing it right now since
we're not actually storing the state of what the "correct" behavior is
that should be changed to, so instead we're just using the issue message
to give us information about what the autocorrect behavior should be.
It's not great, but it works!