Add basics for unused string operation autocorrect
This adds the first bits of an implementation for how we might want to
autocorrect an unused String operation. Unfortunately because there's
nothing in the AST that represents **nothing**, anything we replace that
unused String operation with in the AST will be compiled into something
- even if just the literal atom `nil`. That would probably be better
than unused String operations since those would just be noops in the
runtime, but they would be potentially confusing in the code.

If we want to actually do this there's probably a _lot_ more that would
need to be done for this check, but what we have offers a glimpse into
how this kind of check could be autocorrected as a proof of concept.