Add Credo.Check.Design.AliasOnlyToBoundariesTest module
This is a check which

 * lets you know if you are aliasing too far into other modules
 * lets you know if you should be aliasing deeper into modules

My choosing the level of your alias it is possible to make it clear in your code (i.e. inside of your functions) when a module is a reference to an external or internal module.

Part of this check makes suggestions for aliases similar to the `AliasUsage` check, though `AliasUsage` requires you to alias directly to the module.

This check also allows an `exceptions` option which can be useful for modules which are references very frequently.  For example `MyApp.Blog.Post` may be referenced many, many times and there may only be one module ending in `.Post`, so you can make `MyApp.Blog.Post` an exception so that you could alias directly to it from anywhere.