If reading from standard input, do not consult the filesystem
Files coming in via standard input are not on the filesystem, and
consulting includes and excludes makes an execution run hundreds of
times slower than it should be.

This PR checks to see if the execution is reading from standard input
and if so, returns an empty list to bypass the checks.

Fixes #1054