Simplify SARIF test
This test was running Credo on Credo and asserting what the first result
was, which is not a good test design as completely unrelated changes to
the code can make the SARIF test fail.

The new design is only slightly better, as it is not yet testing the
structure of the result, but at least it is no longer dependent on
everything else in the project.