Make MatchInCondition warn on matches that contain conditionals
This makes MatchInCondition slightly more restrictive in response to a bug we ran into at work.

The problematic code started out looking like this:

```elixir
if user_id = post.user_id do
  do_stuff(user_id, post)
end
```

...but was then modified with the intent that we should `do_stuff` only if we both had a user ID _and_ the post was already published:

```elixir
if user_id = post.user_id && post.published? do
  do_stuff(user_id, post)
end
```

That is, of course, not what the code actually does, but neither I, the author, nor two other reviewers caught that `user_id` was going to get assigned the value `true` when we entered the block.

With this change, `MatchInCondition` will be persnickety about cases like these and require you to extract the match out of the conditional.