Speed up exclusions
On a large project (~3500 files), if I had a check configured with some very simple exclusions like this:

```elixir
{Credo.Check.Warning.ForbiddenModule,
  files: %{
    excluded: [
      "lib/foo.ex",
      "lib/foo_test.exs"
    ]
  }
```

(Note the absence of globs or anything that might need recursive directory traversal)

...I was seeing the check runtime balloon from ~300 ms to nearly 3 seconds.

Upon investigation, I discovered it was because of the `true` condition in Credo.Check.Runner, which looks like this:

```
exec
|> Execution.working_dir()
|> Credo.Sources.find_in_dir(files_included, files_excluded)
```

When the included list was empty, it was doing a full, recursive directory traversersal to find all .ex and .exs files in the project, and this step took about 2.5 seconds on my M4 Max MacBook Pro.

This PR vastly improves that time by providing the list of all (already) known files, and avoiding touching the filesystem at all when the include/exclude paths contain no wildcard characters.

This change brought the overall runtime of Credo on my system (`time mix credo`) from ~17 seconds down to 10 (with Credo's built-in reporting going from "Analysis took 11.1 seconds" to "Analysis took 5.8 seconds"). (The remaining time, of course, is being spent loading the source files into memory initially, prior to any checks.)